Made meeting creation a separate activity
