+++ app/src/main/res/layout/fragment_main.xml
@@ -7,171 +7,16 @@
     tools:context=".MainActivityFragment"
     tools:showIn="@layout/activity_main">
 
-    <!--
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-   -->
-    <TextView
-        android:id="@+id/string_starttime_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="165dp"
-        android:layout_marginTop="28dp"
-        android:layout_marginEnd="154dp"
-        android:layout_marginBottom="12dp"
-        android:text="@string/starttime_label"
-        app:layout_constraintBottom_toTopOf="@+id/date_starttime"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/string_endtime"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="167dp"
-        android:layout_marginEnd="159dp"
-        android:layout_marginBottom="7dp"
-        android:text="@string/endtime_label"
-        app:layout_constraintBottom_toTopOf="@+id/date_endtime"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/date_starttime" />
-
-    <EditText
-        android:id="@+id/date_starttime"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginEnd="85dp"
-        android:layout_marginBottom="16dp"
-        android:ems="10"
-        android:inputType="date"
-        app:layout_constraintBottom_toTopOf="@+id/string_endtime"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/date_endtime"
-        tools:text="@tools:sample/date/hhmm" />
-
-    <EditText
-        android:id="@+id/date_endtime"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginEnd="85dp"
-        android:layout_marginBottom="7dp"
-        android:ems="10"
-        android:inputType="date"
-        app:layout_constraintBottom_toTopOf="@+id/string_members_label"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_endtime"
-        tools:text="@tools:sample/date/hhmm" />
-
-    <TextView
-        android:id="@+id/string_members_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="143dp"
-        android:layout_marginTop="2dp"
-        android:layout_marginEnd="143dp"
-        android:layout_marginBottom="8dp"
-        android:text="@string/members_label"
-        app:layout_constraintBottom_toTopOf="@+id/string_members_input"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/date_endtime" />
-
-    <EditText
-        android:id="@+id/string_members_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="85dp"
-        android:layout_marginBottom="17dp"
-        android:ems="10"
-        android:inputType="textMultiLine"
-        app:layout_constraintBottom_toTopOf="@+id/string_availabilities_label"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_members_label" />
-
-    <TextView
-        android:id="@+id/string_availabilities_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="164dp"
-        android:layout_marginEnd="103dp"
-        android:layout_marginBottom="5dp"
-        android:text="@string/availabilities_label"
-        app:layout_constraintBottom_toTopOf="@+id/string_availabilities_input"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_members_input" />
-
-    <EditText
-        android:id="@+id/string_availabilities_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginTop="3dp"
-        android:layout_marginEnd="85dp"
-        android:layout_marginBottom="9dp"
-        android:ems="10"
-        android:inputType="textMultiLine"
-        app:layout_constraintBottom_toTopOf="@+id/string_location_label"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_availabilities_label" />
-
-    <TextView
-        android:id="@+id/string_location_label"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="162dp"
-        android:layout_marginTop="1dp"
-        android:layout_marginEnd="129dp"
-        android:layout_marginBottom="21dp"
-        android:text="@string/location_label"
-        app:layout_constraintBottom_toTopOf="@+id/string_location_input"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_availabilities_input" />
-
-    <EditText
-        android:id="@+id/string_location_input"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginEnd="85dp"
-        android:layout_marginBottom="66dp"
-        android:ems="10"
-        android:inputType="textMultiLine"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_location_label" />
 
     <Button
-        android:id="@+id/button_create_meeting"
+        android:id="@+id/button_new_meeting"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="143dp"
-        android:layout_marginEnd="107dp"
-        android:layout_marginBottom="17dp"
-        android:text="@string/button_create_meeting"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginStart="8dp"
+        android:layout_marginTop="8dp"
+        android:layout_marginEnd="8dp"
+        android:text="@string/string_new_meeting"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/string_location_input" />
-
+        app:layout_constraintTop_toTopOf="parent" />
 </android.support.constraint.ConstraintLayout>
