+++ app/src/main/res/values/strings.xml
@@ -26,8 +26,6 @@
     <string name="login_error">"Login failed with code %d"</string>
     <string name="login">"Set Sail!"</string>
 
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="action_logout">Log out</string>
     <string name="logout_success">Logged out</string>
     <string name="snackbar">Snackbar</string>
@@ -46,4 +44,6 @@
     <string name="availabilities_label">Availabilities</string>
     <string name="location_label">Location</string>
     <string name="button_create_meeting">Create Parley!</string>
+    <string name="string_new_meeting">Create a parley!</string>
+    <string name="login_successful">Welcome aboard, %1$s!</string>
 </resources>
